final commit! Added POST function to views.py, added CSS styling, added client-side validation and basic client side error handling and updated the classes in rewardsservice to better reflect OOD design and get/post logic
