+++ source/RewardsService/rewardsservice/handlers/customer_handler.py
@@ -3,7 +3,6 @@ import tornado.web
 
 from pymongo import MongoClient
 from tornado.gen import coroutine
-import math
 
 
 #following OOD design best practices, I am creating a class to handle the customer operations and rewards
@@ -14,11 +13,12 @@ class CustomerHandler(tornado.web.RequestHandler):
     def get(self): 
         client = MongoClient("mongodb", 27017)
         db = client["Customers"]
-        email_address = self.get_argument("emailAddress", None)
-        if not email_address:
+        emailAddress = "test2@test.com"
+        emailAddress = self.get_argument("emailAddress", None)
+        if not emailAddress:
             self.write("This email doesn't exist in the system!")
         else: 
-          self.write(db.customers.find_one({"emailAddress": email_address}, {"_id": 0}))
+          self.write(db.customers.find_one({"emailAddress": emailAddress}, {"_id": 0}))
 
 
     # Adding this post method to add an order for an existing customer else create a new customer
@@ -26,14 +26,14 @@ class CustomerHandler(tornado.web.RequestHandler):
     def post(self): 
         client = MongoClient("mongodb", 27017)
         db = client["Customers"]
-        email_address = self.get_argument("emailAddress", None)
-        order_total = self.get_argument("orderTotal", None)
-        if not email_address:
+        emailAddress = self.get_argument("emailAddress", None)
+        orderTotal = self.get_argument("orderTotal", None)
+        if not emailAddress:
             self.write("This email doesn't exist in the system! Adding new customer")
             # insert a new one to the database
-            db.customers.insert({"emailAddress": email_address}, {"rewardPoints": order_total})
+            db.customers.insert({"emailAddress": emailAddress}, {"rewardPoints": orderTotal})
         else: 
-          self.write(db.customers.find_one({"emailAddress": email_address}, {"_id": 0}))
+          self.write(db.customers.find_one({"emailAddress": emailAddress}, {"_id": 0}))
